Example resource: water
This commit is an example of adding a new resource. In this case, water.

Included is a demo map, an entity to gather the new resource from, and a collectable treasure. (Not included are gathering and returning-with-resource animations.)

Icons and cursors are by MuteLovestone, and were copied from https://github.com/0ADMods/resource__water